flaskbook_api/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.4
opencv-python==4.5.1.48
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
regex==2021.11.10
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
typing-extensions==4.0.0
Werkzeug==2.0.2
